tast: Check dependencies before building tests.
When run with -build (currently the default), "tast run ..."
rebuilds the test executable. To do this, it relies on the
executable's dependencies' source code having already been
checked out as a side effect of the dependencies having
previously been built by Portage. (/usr/lib/gopath is used
by default, although this change adds a -sysgopath flag that
can be used to specify a different workspace.)

This change makes tast use Portage's equery command to check
that the test package's dependencies are installed before
performing the build. If they aren't, a command that can be
used to install them is printed. Checking dependencies can
take close to a second; -checkdeps=false can be passed to
skip the check.

BUG=chromium:737628
TEST=unit tests pass; also manually verified that for both
     local and remote tests, an error is printed if
     dependencies are missing and that the check is skipped
     when -checkdeps=false is passed

Change-Id: I532821307e638f27b7dfe013e840b371c63b7547