tast: Use latest system Go workspace when building tests.
When run with -build (currently the default), "tast run ..."
rebuilds the test executable. To do this, it relies on the
executable's dependencies' source code having already been
checked out as a side effect of the dependencies having
previously been built by Portage.

The /usr/lib/gopath directory was previously hardcoded for
this purpose, but this only works if the dependencies were
emerged for the host system. It's more likely that they were
emerged for a board, so scan over /build/*/usr/lib/gopath as
well to find what looks like the most-recently-updated
workspace. This change also adds a -sysgopath flag that can
be used to specify a particular workspace to use.

This is still a bit of a hack, in that there's nothing that
guarantees that the dependencies' source code is up-to-date.
The -build flag is really only useful when developing a
test; though; it's expected that using the test executable
that will be baked into the DUT (via the
chromeos-base/tast-local-tests package) will be the common
case in the future.

BUG=chromium:737628
TEST=tests pass; also manually verified that the latest
     or manually-specified workspace is used by "tast run"

Change-Id: I532821307e638f27b7dfe013e840b371c63b7547