Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > i'm also wondering whether there's some way i can have tests exercise all of the bundle- (and now also runner-)building code in local.go. the main hurdle there is that i don't have a good answer for using the host package's copy-via-tar code in conjunction with the fake SSH server from host/test. i may be able to extend the server to support rewriting tar commands so they're rooted within a temp dir, but that's probably best saved for a different change.
> 
> Does the fake SSH server implement the same interface as the real one? Maybe you could pass the fake implementation to the copy-via-tar code? This is an area where Go lacking generics is a problem: testing with fakes.

i was imprecise in calling it a fake; it's a real ssh server; it's just limited in what it can do. the tests use real ssh client code to talk to the server.

yeah, the hard-to-test-with-fakes thing still irks me a bit. my understanding based on what i've read in the past (at https://github.com/golang/go/wiki/CodeReviewComments#interfaces, i think) is that the official stance is that it's preferred to use real dependencies in tests whenever possible: "Do not define interfaces on the implementor side of an API 'for mocking'; instead, design the API so that it can be tested using the public API of the real implementation."

Patch-set: 2
Reviewer: Dan Erat <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>