tast: Add support for parsing Chrome crash reports.
Chrome's crash-dumping code wraps minidump data in a bespoke
file format that's parsed by crash_reporter. Add a
ReadCrashReport function to cmd/tast/breakpad that can be
used to extract key/value pairs and determine where the
actual minidump data is.

Also add an IsMinidump function that can be used to
determine whether a crash report is a raw minidump file or a
Chrome crash report.

BUG=chromium:770562
TEST=added unit tests

Change-Id: If8e805c8a9a9d772a6c80303c375cfa45384428d