tast: Sort tests by bundle and by test name.
Update test bundles to always run tests in lexicographical
order, and update test runners to always exec test bundles
in lexicographical order.

I think that typically this happened already, but it was
incidental due to the order in which we compile test files
and the implementation of filepath.Glob.

Local test bundles are still all executed before remote test
bundles, and tests from different bundles are not
interleaved.

BUG=chromium:831849
TEST=added unit tests; also verified that tests are still
     run in order

Change-Id: Ib2557d71b93cd695d60ef9b98f66fec2edffc521