tast: Move a few flags out of runCmd.
Move the -checkdeps (renamed to -checkbuilddeps) and
-sysinfo flags from runCmd to the cmd/tast/run package
(that's shared with listCmd). -sysinfo is irrelevant to (and
hidden for) listCmd, but -checkbuilddeps applies to both
the run and list commands. Also stop exporting some
build-related fields from run.Config.

-buildtype still lives in runCmd (with listCmd having its
own -type flag), but this functionality will likely also be
consolidated into the run package in a later change.

BUG=chromium:831849
TEST=unit tests pass; also manually verified that run and
     list commands still work with -build and -build=false
     for local and remote tests

Change-Id: I68d55fa17cce398029d3f3fde2cd47200ce69117