tast: Refactor test-listing code.
Move test-printing functionality out of the cmd/tast/run
package and into listCmd in cmd/tast. cmd/tast/run now takes
a Mode enum that specifies whether it should run or list
tests, and ListTestsMode returns tests via a slice of
TestResult structs, the same as RunTestsMode.

This will make it possible to update the list command to
merge both local and remote tests instead of only listing
one type or the other.

BUG=chromium:831849
TEST=updated unit tests; also verified that "tast run" and
     "tast list" work as before

Change-Id: I7c1a92ef34cc8c4825c7aa00b663e713e2a380fa