tast: Update tast command to check test dependencies.
Update the tast command to use local_test_runner to get the
DUT's software features and to later pass those features
when running tests so that test bundles can skip tests with
unsatisfied dependencies.

Dependency checking is controlled by a new -checktestdeps
flag on the "run" subcommand. The default value, "auto",
attempts to do the right thing (dependencies are checked
when the DUT reports them and an attribute expression was
used to match tests, as happens in the lab). "always" and
"never" values are also available to force specific
behavior.

TestResult objects written to results.json are updated to
include a new "skipReason" string field that is nonempty if
the test was skipped (currently, just due to missing
dependencies).

BUG=chromium:835991
TEST=added unit tests; also manually tested by manipulating
     /etc/tast_use_flags.txt, tests' SoftwareDeps fields,
     and the -checktestdeps flag and test patterns passed to
     the "tast run" command

Change-Id: If4f8bfcb14e91f845c38081f4b58570011fcc91f
Reviewed-on: https://chromium-review.googlesource.com/1038892
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>