tast: Disallow tests with negative timeouts.
Make testing.Registry.AddTest return an error when passed a
test with a negative timeout. If anyone did this, it could
cause problems once the tast.py Autotest test uses test
timeouts as a guide when choosing a timeout for the "tast
run" command.

BUG=chromium:851572
TEST=updated unit tests

Change-Id: I94c00355cf1fee7b13d8d19a0415124729d0f4c9
Reviewed-on: https://chromium-review.googlesource.com/1096367
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>