tast: Add -timeout flag to run subcommand.
Add a -timeout flag to the tast executable's "run"
subcommand that can be used to limit the total running time.
The tast Autotest test can set this to make it more likely
that the tast process will detect an impending timeout and
have time to write its results to disk before it's killed.

Also watch the context's deadline while reading control
messages.

Finally, check the SSH connection at completion and log an
error if it's no longer active.

BUG=chromium:778389
TEST=added and updated unit tests; also manually tested with
     various timeouts

Change-Id: I3fe99324bcc6ed41b19c83d6ef23b71b712ca308