tast: Make host package honor context deadlines.
Update the SSH type's Start, PutTree, and GetFile methods
and SSHCommandHandle's Close and Wait methods to honor
context deadlines.

This makes it easier to guard against DUTs that hang
mid-test; for instance, the tast process can detect the
issue and exit with a friendlier error message instead of
hanging indefinitely.

BUG=chromium:778389
TEST=added unit tests

Change-Id: Icd4656fffa31e80f43c8a34c95ac742a6964a120