tast: Avoid duplicating bundle and runner args.
runner.Args was declaring its own RemoteArgs and RunTestArgs
structs that were identical to the ones in bundle.Args and
were just copied into it.

Simplify the code by embedding bundle.RemoteArgs in
runner.RemoteArgs and bundle.RunTestArgs in
runner.RunTestsArgs. (Keeping the runner structs makes it
possible to add additional runner-specific args in the
future.)

Both the bundle and runner structs are
marshaled to JSON in the same way, so this retains backward
compatibility.

BUG=none
TEST=unit tests pass

Change-Id: I8fad3becd3c5311dad4ea12d5bb6092555d0c090
Reviewed-on: https://chromium-review.googlesource.com/1143012
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>