tast: Remove 'lg' logging.Logger global.
Remove the tast command's global 'lg' logging.Logger
variable. Instead, attach a logger to the context that is
passed to subcommands using added NewContext and FromContext
functions in the logging package.

The logger unfortunately can't be easily assigned directly
to fields in subcommand structs, as subcommands must be
created before parsing flags, while the logger can only be
created after doing so.

I'm doing this to make it possible for subcommand tests to
verify log messages in a later change. This change also adds
a logging.NewDiscard convenience function for tests that
don't care about logging.

BUG=chromium:778389
TEST=unit tests pass and tast command logs output as before

Change-Id: If9ea532755711d230e18d1019974ee5fce9e2c3a