tast: Update test.SSHServer to expose delays for fake cmds.
Update the SSHServer type's FakeCmd method to take an
exported CmdResult struct that contains optional StartDelay
and EndDelay fields. These are equivalent to the delays that
can be set via the existing ExecDelays method, but allow
delays to be configured for individual commands. The delays
configurable by ExecDelays are updated to actually-executed
commands.

BUG=chromium:778389
TEST=unit tests still pass

Change-Id: I861b1d4e97c18bd7126a2f7f7b9dcb4d9166879c
Reviewed-on: https://chromium-review.googlesource.com/1154492
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>