Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Any chance this could hang forever?

in unit tests, you mean, or in general?

when tast is running in the lab, there's a deadline set on ctx, so we'll report failure if the command is hanging on the DUT.

in unit tests, we're connecting to an SSH server on the local machine that's been configured to block for 30 seconds, so we shouldn't hang there either.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>