tast: Check data files are declared.
Make DataPath to check if a given file name is declared in Test.
Also, check the validity of file names in AddTest to catch errors
early.

This will avoid confusing "no such file or directory error" when
adding a new test because undeclared data files are not pushed to
DUT.

BUG=chromium:873952
TEST=fast_build.sh -T
TEST=tast run DUT example.DataFiles

Change-Id: Ica9e4ddfba7d4c9acfb184e2834952361ca8d7e8