tast: Rework and simplify build flags.
Try to improve the way that build parameters are passed to
the tast command. Previously, there was no way to specify an
additional location containing the test bundle code while
still including src/platform/tast-tests (which contains
packages like chrome and arc that will likely be necessary
to many non-public tests).

Remove the build package's flags (-buildarch, -buildoutdir,
-buildsysdir, -buildtestdir, -buildcommondir) and replace
build.Config's TestWorkspace, CommonWorkspace, and SysGopath
fields with a single Workspaces field that callers can set
to provide the location of the source code to be built.

Define -buildworkspace and -buildoutdir flags in the run
package instead. -buildworkspace can be used to specify an
additional workspace that should be used when building test
bundles.

Also give the build subcommand its own -arch and -workspaces
flags.

BUG=chromium:875141
TEST=verified that -build=true still works for local and
     remote tests. -buildlocaltestrunner still works. also
     created a new workspace containing an "ext" test bundle
     and was able to build and run it with the following:
     tast run -build=true -buildbundle=ext \
       -buildworkspace=/tmp/workspace \
       -checkbuilddeps=false <dut>

Change-Id: I6ef0373e222d34a2c3a255f79ee4a40ec115a743