tast: Relax func/file name rules for acronyms like 'WebRTC'
This commit relaxes the rule for function/file names to deal with
acronyms where uppercase letter apperar aftert lowercase letter
as one word in filename.

For this purpose, checkFuncNameAgainstFilename doesn't care if
'_' exists between words in filename, because we cannot know if a
given string is one word or not.

BUG=chromium:880661
TEST=~/trunk/src/platform/tast/fast_build.sh -T

Change-Id: I887c83e7c58af6db3ef1f2aee52c12e6e1bff7c3
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>