tast: Add -extrauseflags to tast and local_test_runner.
Add a -extrauseflags flag to the "tast run" subcommand that
can be used to inject a comma-separated list of additional
USE flags to treat as being present when determining a DUT's
features for the purpose of checking test dependencies.

Also add an identically-named flag to local_test_runner, and
introduce a new command.ListFlag type for splitting flag
values into slices.

This can be used to exercise dependency-checking code, and
it can also be set by VM builders that test non-VM-specific
images like amd64-generic. Right now, we can't run Tast
tests on these builders since tests that depend on non-VM
features like audio hardware and TPMs run and fail. With
this, we can get the same behavior that we already get when
testing dedicated VM images like betty.

BUG=chromium:885016
TEST=added unit tests; also ran "tast run" with
     -extrauseflags=chromeless_tty and -checktestdeps=always
     and verified that ui.ChromeLogin gets skipped, and that
     local_test_runner behaves similarly with -extrauseflags

Change-Id: I427df7b7b60591f522d7fe294958ff95d5fd05a6