tast: Add support for test preconditions.
Add a Pre field to the testing package's Test struct that
can be used to specify a precondition that must be satisfied
before the test is run.

Preconditions (represented by a new testing.Precondition
interface) contain setup code to get the DUT into a state
like "Chrome is logged in" or "ARC has booted". Performing
this setup once and then running all tests that require it
can greatly speed up testing.

This change also modifies the order in which tests are run:
tests are grouped by precondition before being sorted
lexicographically.

BUG=chromium:892009
TEST=added/updated unit tests; also manually verified in
     conjunction with a change in tast-tests that defines
     and uses preconditions for logging into chrome and
     booting android

Change-Id: I71975d15ff0d8594243065800fe952419443f22b