tast: Make Test.Run construct its own State.
Introduce a new TestConfig struct in the testing package and
pass it to Test.Run instead of passing a State struct.

This improves encapsulation (there was no reason that the
bundle package needed to know about testing.State) and
readability (the NewState function had too many arguments).

It also makes it possible for future changes to refactor
Test.Run to e.g. perform additional preliminary work without
it counting against the test's timeout.

BUG=chromium:890546
TEST=updated unit tests; also manually verified that local
     and remote tests still work

Change-Id: I94c8a528040d3b71dfa04aa3f8983c17199496c1