tast: Allow defining custom error types.
Currently, we do not have a way to define custom error types that
supports error chaining and stack traces. While there is no use
case as of today, we are afraid that lack of custom error type
support can discourage users to inspect errors in proper ways.

This CL make it possible to define custom error types by embedding
*E.

BUG=chromium:887166
TEST=fast_build.sh -T -C

Change-Id: Ia221b3bdff631f20d513ac0d190900d922055f60