tast: Refactor Test.Run to use stages.
Generalize the testing package's code for running a test
into a runStages function that runs a series of functions
(currently setup, test, cleanup) within a goroutine. Each
function has its own timeout, which is a departure from the
old model of also allocating the test's timeout to setup and
cleanup. This paves the way for granting longer timeouts to
preconditions in the future.

The "tast run" command limits the amount of time that it
waits for the next control message (to avoid hanging
indefinitely if the DUT becomes unresponsive), so this
change also adds a new Test.AdditionalTime field that the
bundle uses to pass back extra time allocated to setup and
cleanup functions.

BUG=chromium:892009
TEST=added unit tests for runStages; also verified that
     tests still work and that "tast list -json" shows that
     additionalTime field is present in serialized tests

Change-Id: I93f0f77159c5847a8b055db4d40b6104be9e4965