Update patch set 2
Patch Set 2:

thanks for doing this, and nice work! when i saw the change in my inbox i was hoping this was using go's parsing packages. :-)

this seems like it will give us a lot of flexibility to detect different issues. i find that i write a lot of the same code review comments over and over as people send their first test changes, so it'll be nice to try to automate some of that in pre-upload hooks.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>