tast: Add outputReader for unit tests in testing package.
Add a new outputReader struct that implements an
infinitely-buffered chan Output.

This prevents an issue that I frequently encounter when
making changes to the testing package. When additional
Output messages are written when running tests, unit tests
hang because the number of messages that are written now
exceed the buffering of the output channel.

BUG=chromium:892009
TEST=unit tests still pass

Change-Id: I5d51981f5539c9530906304eacc4247eb5d1c3f5