tast: Make expr package support wildcards.
Update the expr package (used to evaluate boolean
expressions specifying which tests to run or how to
determine features from USE flags) to treat '*' characters
in identifiers as wildcards.

I've heard requests in the past to have some way to match
all tests named "ui.*" in attribute expressions, and this
also seems like it will be helpful if we want to add
test attributes like "suite:foo" and then exclude those
tests from the builders in the CQ and PFQs.

BUG=chromium:894375
TEST=added unit tests

Change-Id: I1f7eba748580d3a339e022c5ee6d6433d0699ca6