tast: Add Tast logo and embed it in README.md.
Also add logo.svg, a simplified version of the original
Inkscape SVG file (i.e. with some temporary layers deleted).
Note that the text kerning is wrong when viewed in Chrome;
the file should be loaded in Inkscape and exported to PNG
from there if a higher-resolution bitmap is needed.

BUG=none
TEST=viewed in md_browser.py

Change-Id: I262bac66e4464748b5fbbd35ac7ff1d87054f2f6