tast-lint: Run Golint.
We run golint and take problems whose confidence is >=0.8.

We filter results in addition in some ways:
- Lines marked "// NOLINT" are ignored. This is because golint is
  actually designed to make "suggestions"; it sometimes gives
  false positive but does not provide a way to selectively ignore
  errors by itself. Nevertheless, we believe running golint in
  the pre-upload hook is worth doing.
- unexported-type-in-api is ignored.
- Test main functions are allowed to be exported without
  comments.

Also, -debug option is added to turn on verbose output. It can be
used to print the information required to add more filtering
in the case of false positives.

CQ-DEPEND=CL:1287349
CQ-DEPEND=CL:1287449
BUG=chromium:894750
TEST=fast_build.sh -T
TEST=tools/run_lint.sh

Change-Id: Ic095877d0069ca917e9fb406b087cf458bcac78d