tast-lint: Checks inter-file references in test main files.
TestMain check is removed and two new checks are added instead:

1. InterFileRefs: checks that test main files do not have inter-file
   references.
2. Exports: checks that test main files does not export anything but
   test main functions.

Introduction of the InterFileRefs check finally allows to loosen the
current unideal restriction that test main files should not have any
top-level declarations but test main functions.

In order to check inter-file references, we have to parse other Go
files in the same directory to build the complete map of declared
objects in a package. cachedParser is introduced for this purpose.

BUG=chromium:882022
TEST=fast_build.sh -T
TEST=run_lint.sh src/**/*.go

Change-Id: I52f2ee33423d4513eaad5a8349bea6cd905b5bb6
Reviewed-on: https://chromium-review.googlesource.com/1297560
Commit-Ready: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>