tast: Add "no_android" software feature.
Add a new "no_android" software feature that's the opposite
of the existing "android" feature. This allows us to have
two versions of a test, one that runs on Android-supporting
devices and one that runs on devices that don't support
Android.

BUG=chromium:877733
TEST=none

Change-Id: I5f15b434ff85906e15b4b4f463c98da385778c86
Reviewed-on: https://chromium-review.googlesource.com/1342947
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>