tast: Add "reboot" software dependency.
Add a new "reboot" software feature that remote tests that
reboot the DUT can depend on. This is really just a hack to
let us skip power.Reboot from running on VM builders, where
we're apparently sometimes seeing the local_tast_runner
executable go missing after a reboot. The cause of this is
currently unknown.

BUG=chromium:909955
TEST=none

Change-Id: Ib266ac964075a106547e909f232a22156fb080bb