tast: Add -waituntilready flag to "run" subcommand.
Add a -waituntilready flag to "tast run" that controls
whether the test bundle's "ready" function (if any) will be
executed before running tests. Before this change, the
function was always executed.

The local "cros" bundle's readiness function waits for
various Upstart services, which is low-cost, but it also
attempts to check the TPM's ownership state and initialize
it, which can be slow or can fail entirely.

This flag currently defaults to true, but I'm planning to
make it default to false after updating builders to
explicitly set it.

BUG=chromium:914258
TEST=updated unit tests; also manually verified that
     readiness is checked by default for "tast run" and
     local_test_runner but not checked when passing
     -waituntilready=false to either command

Change-Id: I0c9d1e39437d270204df0f693d5655052d8ea903
Reviewed-on: https://chromium-review.googlesource.com/1375010
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>