Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I'm afraid --onlydeps is not enough to update dependencies. In fact I saw the following situation:
> > 
> >  $ tast run localhost:9222 '(!disabled && "name:arc.*")'
> >  ...
> >  2018/12/18 18:35:12 Failed building or pushing tests: build failed: chromeos-base/tast-local-tests-cros-9999 has missing dependencies
> >  To install missing dependencies, run:
> >  
> >   sudo emerge -j 16 \
> >     =chromeos-base/policy-go-proto-0.0.1-r5
> > 
> >  $ sudo emerge --onlydeps chromeos-base/tast-local-tests-cros
> >  Calculating dependencies... done!
> >  >>> Auto-cleaning packages...
> > 
> >  >>> No outdated packages were found on your system.
> 
> is this maybe caused by not having tast-local-tests-cros cros_workon-ed in the host sysroot? (i'm not sure offhand if emerge can take a version number, either.)

It looks like this is caused because the packages were already previously installed as an older version, so in this case, something like 'chromeos-base/policy-go-proto-0.0.1-r4' was already installed, so the emerge command I gave would see that the r4 revision of that package was already installed, so no further action would need to be taken.

To resolve this, we just need to add the --update (-u for short) and --deep (-D for short) flags to the emerge command so it can rebuild the package and its dependencies if there's a newer revision detected (to keep the command short, we can shorten --onlydeps to -o).

The new command would look like:
sudo emerge -j 16 -uDo chromeos-base/tast-local-tests-cros

Patch-set: 1
Reviewer: Gerrit User 1155964 <1155964@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>