Update patch set 4
Patch Set 4:

Hmm, having three different flags for proxy configuration sounds a bit involved...

There are two cases about proxy setup: (1) the same proxy settings should be used on host and DUT, and (2) different proxies should be used. It seems like (1) is the case with both Intel and Google.

The default of the current patch works for (1), which is great. And I guess the three flags were defined to support any case of (2). But if proxy is required for host but not for DUT, we need to specify three flags with empty value (-httpproxy= -httpsproxy= -noproxy=). This suggests me that it might be better not to take the default values for the flags from environment variables. Hmm...

Now I feel it's okay to not support (2) and always use the same proxy configuration on host and DUT for now because there's no real use case yet (sorry for reversing my last comment).

Patch-set: 4
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>