Update patch set 3
Patch Set 3:

> apologies for my slow reply here. i think that we don't want to specify exact version numbers in the tast-local-tests-cros ebuild. that seems like a maintenance nightmare, and as far as i'm aware we don't do it in many other chrome-os-maintained ebuilds.
> 
> i'm still slightly anxious about the provided command upgrading a set of packages that don't exactly match those that are identified by the dependency check, but as long as the former is a superset of the latter rather than a subset, i think it's okay.
> 
> in a perfect world, "tast run -build=true" would just emerge all of tast-local-tests-cros's outdated/missing dependencies every time it wants to build the bundle, but i don't think there's any way to do that without substantially degrading the developer experience due to how slow emerge is when checking dependencies. as a compromise, i settled on the current (still slower than i'd like) approach of a "equery g" command, followed by a bunch of "equery l" commands in parallel, followed by prompting the dev to manually emerge outdated direct dependencies. the intent is that devs just need to run it once after syncing their checkout.

Apologies for replying late myself, I've been working on and off over the past two weeks on a better solution for the dependencies, but ultimately didn't like what I came up with.

My solution for what might work well in terms of speed and ease of use is to change the dependency checks to use a virtual ebuild, like Luigi suggested in the cros-dev thread, so we can only install the dependencies and not worry about the other modifications made in the tast-local-tests-cros ebuild.

Then, if we use 'emerge --search <package>', we can just use some regex checks to confirm that the installed version matches the latest version available (in case an update to the dependencies is required).

The pros of this method would be:
- Single-line command for updating (ie: sudo emerge -j 16 tast-local-cros-deps).
- Only need to check a single package to ensure it's installed and up to date instead of looping over dependencies to ensure they're installed.

Cons:
- Need to maintain the virtual ebuild to ensure it stays in sync with the dependencies from tast-local-tests-cros.
- Dependencies won't be have a version requirement unless specified in the virtual ebuild (will not always be latest and greatest).

If this sort of method is undesirable, we can use the single-line method I've suggested in this change (I'll update it with go fmt first) or continue to use the current dependency checks. Just looking to provide some options here :)

Patch-set: 3
Reviewer: Gerrit User 1155964 <1155964@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>