tast: Support 64bit ARM userland.
Formerly, we always assumed 32bit userland on 64bit CPUs. However
we are introducing 64bit userland on some boards (e.g. kevin64).
This difference is critical on selecting the Go toolchain to
build binaries since we do not support running 32bit binaries on
64bit userland.

This change upates the functions to detect the target architecture
to get the userland architecture, instead of the kernel architecture,
and corrects the toolchain selection.

To get the userland architecture, we run the "file" command against
an arbitrary binary on the target (/sbin/init for now), instead of
"uname -m". I'm skeptical if it's an optimal way to check the
architecture, but I could not find a better way.

BUG=chromium:918245
TEST=tast run DUT platform.P2PClient  # w/ betty, kevin, kevin64

Change-Id: I8e036200a6cf820ff53d13252d7f0ca277f4cc78