Update patch set 4
Patch Set 4:

> if this is changed, my preference is for the main focus to be on making it so we don't check (or emerge) RDEPEND packages. those are unnecessary for building test bundles, and avoiding them would be a much bigger win than having a shorter command line.
> 
> is there some way to make portage either print a package's direct build (but not runtime) dependencies, or to at least evaluate the ebuild file (and its inherits) so we can get the full DEPEND list?

It appears that:
$ emerge --onlydeps --onlydeps-with-rdeps=n -pv tast-local-tests-cros'
now shows the DEPEND dependencies from the ebuild that we want (I assume this now works because of the Portage updates that Mike was working on recently). \0/

So, all we'd have to do is add the '--onlydeps-with-rdeps=n' --onlydeps option to the one-line command.

Patch-set: 4
Reviewer: Gerrit User 1155964 <1155964@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>