Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> PTAL
> 
> ISTM we have a -9999 ebuild covering this. Will I need to wait for a revbump to land on the corresponding -rNNN ebuild before submitting https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1413873 ?
> 
> Thanks!

you should add a CQ-DEPEND line on this change that references the tast-use-flags change, and a CQ-DEPEND line on the tast-tests change that references this change.

please also add a description of the new flag to the test_dependencies.md doc in this same repo. thanks!

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>