tast: Update timing.Log to support JSON and importing.
Update the timing package to support marshaling its structs
to JSON and importing one Log into another. This will be
used by a later change to make it possible for test-reported
timing data to be incorporated into the main timing.json
file.

Also add an Empty method that will be used later and update
Stage.End to recursively close any child stages that are
still active.

BUG=chromium:888295
TEST=added unit tests

Change-Id: I2f32e2fbcc7f0860b8d017b21925b3faf448d9e8