tast: Let tests report timing data.
Attach a timing.Log to the context passed to each test that
can be used to report per-test timing data. These logs are
included in control.TestEnd messages and imported into the
main log that is written to timing.json.

The timing package is moved from cmd/tast/timing to
tast/timing to make it available to the bundle package.

BUG=chromium:888295
TEST=updated unit tests; also manually tested in conjunction
     with change to make chrome package report timing stages

Change-Id: Iea04eb1857192d65f54fef190d5f77970fa0d570