tast: Support retrying SSH connection attempts.
Add an -sshretries flag to the "run" and "list" subcommands
that can be used to specify the number of times to retry SSH
connection attempts. Also add a ConnectRetries field to
host.SSHOptions and add a RejectConns method to
test.SSHServer that can be used to simulate failures.

The flag defaults to 0 since I think that that's what most
developers will want, but I plan to set it to 2 when Tast is
run in the lab to see if it helps with connection errors
(presumably caused by network flakiness) that have been seen
when collecting DUT system information in the lab.

Also attempt to resolve a long-standing TODO about SSH
connections being leaked in the unlikely event that they
take longer than the context's deadline but still succeed
eventually.

BUG=chromium:928429
TEST=added a unit test for host package; also manually
     verified that we retry when i unplug my DUT's ethernet
     and pass -sshretries to "tast run"

Change-Id: I5bad8f83f4d252a664203b4d80d8558a067eb1fc
Reviewed-on: https://chromium-review.googlesource.com/1459029
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>