tast: Changing diff to always compare files as text
In the case that the contents from binary files are being compared, the
output from the diff command does not match what is expected and will
return an error. This change adds an extra argument to the diff
invocation to compare all files as if they were text.

BUG=None
TEST=Passes unit tests

Change-Id: I3321708fd64711a7433b4f94669742c0791f96ab
Reviewed-on: https://chromium-review.googlesource.com/1474626
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>