tast: Fix control.TestEnd.TimingLog JSON name.
I incorrectly made control.TestEnd.TimingLog get marshaled
as "timingLog" rather than the correctly-namespaced
"testEndTimingLog". This doesn't break anything right now,
but it would cause issues if we added another TimingLog
field to a different control message type in the future.

Update the TimingLog field to be marshaled as
"testEndTimingLog" and add a TimingLogDeprecated field that
still using "timingLog". To ensure compatibility in both
directions, update WriteMessage to fill TimingLogDeprecated
and ReadMessage to fill TimingLog.

BUG=chromium:932307
TEST=manual: executed old test bundle with new tast command
     and new test bundle with old tast command and verified
     that test-reported timing events were written to
     timing.json

Change-Id: I6ffe96c48261ccab9bc884ae46207504bb95c21c