tast: Avoid writing results dir for compilation errors.
3bf3b54a83fe went too far in making "tast run" write a
results dir regardless of whether any tests are successfully
executed or not. It's undesirable to write results when we
fail to compile tests with -build=true, as this can leave a
bunch of useless results dirs lying around.

Instead, only write results if we at least got to the point
where we tried to run tests. We avoid writing results if we
fail earlier, e.g. while compiling/pushing tests, collecting
initial system information, or collecting the DUT's software
features.

BUG=chromium:928445
TEST=added unit tests; also manually verified that results
     are collected when tests are executed successfully and
     when the ready function fails while -waituntilready is
     used, but not when compilation fails while using
     -build=true

Change-Id: I1965f7e210f382b77fbfb0fd98e3178ce95c5f51