tast: Log errors when collecting system info from DUT.
Make "tast run" log errors (e.g. SSH connection problems)
immediately when it encounters them while trying to collect
system logs and crashes from the DUT.

Formerly, we just reported the error at the end of the log
file, which was confusing. In particular, it could make it
look like the SSH connection attempt succeeded when it
actually didn't.

BUG=chromium:937913
TEST=added a unit test; also forced a timeout while
     collecting system info and verified that the error was
     logged immediately:
     2019/03/04 13:30:37 Ran 1 remote test(s) in 8.09s
     2019/03/04 13:30:37 Collecting system information
     2019/03/04 13:30:42 Connecting to root@...
     2019/03/04 13:30:52 Failed collecting system info: dial tcp ...:22: i/o timeout

Change-Id: Ie0a1f8537983bbb90cf29581d210c9945215d50e